Move schemas out to new submodule
